revert double to real conversion of hbond polys (#369)
Reversion of double->Real conversion of hbond polys (and reversion of unit test changes)

Runtime of this version:
```
  structure    nres   t_tmol(s)    t_ros(s)   E_tmol_start         E_tmol     E_tmol_ros    E_ros_start          E_ros
  ---------  ------  ----------  ----------  -------------  -------------  -------------  -------------  -------------
  1INP          400        7.36      146.55       1307.469      -1030.550       -618.331       1703.449       -806.548
```
Versus float:
```
  structure    nres   t_tmol(s)    t_ros(s)   E_tmol_start         E_tmol     E_tmol_ros    E_ros_start          E_ros
  ---------  ------  ----------  ----------  -------------  -------------  -------------  -------------  -------------
  1INP          400        6.31      110.70       1309.654       -985.013       -580.288       1703.449       -799.147
```

+++ tmol/score/hbond/hbond_energy_term.py
@@ -24,6 +24,8 @@ class HBondEnergyTerm(AtomTypeDependentTerm, HBondDependentTerm):
         # Cache of parameter tables converted to whichever coord dtype is used
         # in scoring. Avoids a per-forward .to(coords_dtype) on three tensors.
         # Keyed by dtype; stored value is a (pair_param, pair_poly, global_param) tuple.
+        # pair_poly_table is always kept at float64 — the C++ kernel templates
+        # HBondPolynomials on double independently of the Real coord dtype.
         self._param_tables_by_dtype: dict = {}
 
     def _param_tables(self, coords_dtype):
@@ -31,7 +33,7 @@ class HBondEnergyTerm(AtomTypeDependentTerm, HBondDependentTerm):
         if tables is None:
             tables = (
                 self.hb_param_db.pair_param_table.to(coords_dtype),
-                self.hb_param_db.pair_poly_table.to(coords_dtype),
+                self.hb_param_db.pair_poly_table.to(torch.float64),
                 self.hb_param_db.global_param_table.to(coords_dtype),
             )
             self._param_tables_by_dtype[coords_dtype] = tables
